backened/requirements.txt
flask
face_recognition_models
dlib
numpy
Pillow
